+++ AdvancedOpenGL/src/engine/Window.h
@@ -15,7 +15,7 @@ public:
     virtual void logGlParams() = 0;
     //virtual bool should_close() = 0;
     //virtual void handle_close() = 0;
-    virtual void updateFpsCounter(long dt) = 0;
+    virtual void updateFpsCounter(float dt) = 0;
     virtual void clear() = 0;
     virtual void swapBuffer() = 0;
     virtual void clean() = 0;
