+++ AdvancedOpenGL/src/game/Scene_001_Color.h
@@ -18,7 +18,7 @@ public:
     void pause();
     void resume();
     void handleEvent(const InputState &);
-    void update(unsigned int dt);
+    void update(float dt);
     void draw();
     void setGame(Game *);
